lua-simple-uci: delete option when using Cursor:set() with nil value
Setting the section type when using set() with only two arguments is a very
confusing behaviour. section() can be used without a value table for the
same result.

Deleting the option instead will make set() and get() symmetric and will
allow to simplify some code that needed to check for nil to call set() or
delete() before.