update tunneldigger files to add hotplug functionality
An issue arrises when a router is not connected to a wan and only meshes.
The wan is not connected to anything, yet tunneldigger tries to make a
connection anyhow.  To aviod this, the init script has been modified
to _not_ make tunnels when they are configured with the bind_interface
option.  Additionally, a hotplug script has been created to create
and destroy tunnels on demand as the interfaces go up and down.  The
hoplug script only works with tunnels configured with the bind_interface
option.

More info at https://github.com/freifunk-berlin/firmware/issues/606

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>