respondd: add provider "respondd-providers" to list registrations
Add an internal respondd provider "respondd-providers" which allows a
respondd querier to query which providers are registered on a node.
Example output:

{
	"respondd-providers": [
		"neighbours",
		"nodeinfo",
		"statistics",
		"statistics-extended"
	]
}

This can be useful for providers which are known to be bulky, where the
actual data is better queried via HTTP/TCP instead. But where such a
provider is only present on a few nodes, so you'd want to avoid
trying each node via HTTP.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>