autoupdater: update download progress display less frequently
Updating the "XXXXX / XXXXX KiB" display for every run of
recv_image_cb() is a significant bottleneck on slow consoles. This was
reported for a device using a 9600 Baud serial console, but the slowdown
is noticeable even on a 115200 Baud console, especially when the network
is fast.

Replace the output with a "XX.X / XX.X MiB" display and only update it
every 0.1 MiB to fix the issue.

Closes #273