ebtables-tiny: reduce wait on missisng lock
With the fix form Shoragan to actually wait if a lock-file is present,
retries would only be done every second.
But: Interacting with ebtables should not take that long.
So we can drastically reduce the waiting time here.

With this change will will only wait 50ms (instead of 1s).
To not spam the log too much if a lock is kept for longer (or the other
process has more luck locking) this change also reduces the logging
rate.

Signed-off-by: Chris Fiege <chris@tinyhost.de>